Conflicts:
	src/server/game/Battlegrounds/Battleground.cpp
	src/server/game/Entities/Pet/Pet.cpp
	src/server/game/Entities/Player/Player.cpp
	src/server/game/Entities/Transport/Transport.cpp
	src/server/game/Entities/Unit/Unit.cpp
	src/server/game/Entities/Unit/Unit.h
	src/server/game/Server/WorldSocket.cpp
	src/server/game/Spells/Auras/SpellAuraEffects.cpp
	src/server/game/Spells/Auras/SpellAuras.cpp
	src/server/game/Spells/Spell.cpp
	src/server/game/Spells/SpellEffects.cpp
	src/server/game/Spells/SpellInfo.cpp
	src/server/game/Spells/SpellMgr.cpp
	src/server/game/Spells/SpellScript.cpp
	src/server/scripts/EasternKingdoms/undercity.cpp
	src/server/scripts/Kalimdor/moonglade.cpp
	src/server/scripts/Northrend/CrusadersColiseum/TrialOfTheCrusader/boss_twin_valkyr.cpp
	src/server/scripts/Northrend/FrozenHalls/PitOfSaron/boss_forgemaster_garfrost.cpp
	src/server/scripts/Northrend/FrozenHalls/PitOfSaron/pit_of_saron.cpp
	src/server/scripts/Northrend/FrozenHalls/PitOfSaron/pit_of_saron.h
	src/server/scripts/Northrend/IcecrownCitadel/boss_blood_prince_council.cpp
	src/server/scripts/Northrend/IcecrownCitadel/boss_blood_queen_lana_thel.cpp
	src/server/scripts/Northrend/IcecrownCitadel/boss_deathbringer_saurfang.cpp
	src/server/scripts/Northrend/IcecrownCitadel/boss_professor_putricide.cpp
	src/server/scripts/Northrend/IcecrownCitadel/boss_sindragosa.cpp
	src/server/scripts/Northrend/IcecrownCitadel/instance_icecrown_citadel.cpp
	src/server/scripts/Northrend/Naxxramas/boss_gothik.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_algalon.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_auriaya.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_flame_leviathan.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_general_vezax.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_xt002.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/boss_yoggsaron.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/instance_ulduar.cpp
	src/server/scripts/Northrend/Ulduar/Ulduar/ulduar.h
	src/server/scripts/Northrend/UtgardeKeep/UtgardePinnacle/boss_svala.cpp
	src/server/scripts/Northrend/dragonblight.cpp
	src/server/scripts/Northrend/sholazar_basin.cpp
	src/server/scripts/Spells/spell_generic.cpp
	src/server/scripts/Spells/spell_priest.cpp
	src/server/scripts/World/npcs_special.cpp

+++ src/server/game/Maps/Map.cpp
@@ -814,20 +814,20 @@ void Map::RemoveCreatureFromMoveList(Creature* c)
 void Map::MoveAllCreaturesInMoveList()
 {
     _creatureToMoveLock = true;
-    for(std::vector<Creature*>::iterator itr = _creaturesToMove.begin(); itr != _creaturesToMove.end(); ++itr)
+    for (std::vector<Creature*>::iterator itr = _creaturesToMove.begin(); itr != _creaturesToMove.end(); ++itr)
     {
         Creature* c = *itr;
-        if(c->FindMap() != this) //pet is teleported to another map
+        if (c->FindMap() != this) //pet is teleported to another map
             continue;
 
-        if(c->_moveState != CREATURE_CELL_MOVE_ACTIVE)
+        if (c->_moveState != CREATURE_CELL_MOVE_ACTIVE)
         {
             c->_moveState = CREATURE_CELL_MOVE_NONE;
             continue;
         }
 
         c->_moveState = CREATURE_CELL_MOVE_NONE;
-        if(!c->IsInWorld())
+        if (!c->IsInWorld())
             continue;
 
         // do move or do move to respawn or remove creature if previous all fail
@@ -2497,7 +2497,7 @@ bool InstanceMap::AddPlayerToMap(Player* player)
                 if (uint32 dungeonId = sLFGMgr->GetDungeon(group->GetGUID(), true))
                     if (LFGDungeonEntry const* dungeon = sLFGDungeonStore.LookupEntry(dungeonId))
                         if (LFGDungeonEntry const* randomDungeon = sLFGDungeonStore.LookupEntry(*(sLFGMgr->GetSelectedDungeons(player->GetGUID()).begin())))
-                            if (uint32(dungeon->map) == GetId() && dungeon->difficulty == GetDifficulty() && randomDungeon->type == LFG_TYPE_RANDOM)
+                            if (uint32(dungeon->map) == GetId() && dungeon->difficulty == uint32(GetDifficulty()) && randomDungeon->type == uint32(LFG_TYPE_RANDOM))
                                 player->CastSpell(player, LFG_SPELL_LUCK_OF_THE_DRAW, true);
         }
